Adds new sample for github webhook integration
Add code and run script.

Add code for an adapter binary to translate github
event payload to a  payload that EN can understand.

Add code for a builder to allow the creation of
a BuildRun from a predefined Build. It will also
modify any existing application referencing the
rebuilded image.

Add run.sh script to automate the deployment of this
architecture.

Signed-off-by: encalada <encalada@de.ibm.com>